pyproject.toml
[project]
name = "classact"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "spotipy>=2.24.0",
    "xgboost>=2.1.1",
    "tqdm>=4.66.5",
    "pyyaml>=6.0.2",
]
